backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
scikit-learn
pandas
numpy
networkx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery
redis
supabase
python-dotenv
